% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstap}
\alias{bootstap}
\title{Runs a parameteric bootstrapping likelihood ratio test on two DAISIE models}
\usage{
bootstap(data, model_1, model_2, seed, cond)
}
\arguments{
\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}} and friends.}

\item{model_1}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_analysis]{run_analysis()}}.}

\item{model_2}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_analysis]{run_analysis()}}.}

\item{seed}{A single numeric to be used as seed.}

\item{cond}{An integer specfying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}
}
\value{
Nothing. Writes bootstrapping results to a \code{.rds} file. This
file is stored in \code{file_path}. The directory in of \code{file_path} is created
if it doesn't exist.
}
\description{
Runs a parameteric bootstrapping likelihood ratio test on two DAISIE models
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
bootstrap(
  data = Galapagos_datalist,
  model_1 = cr_dd,
  model_2 = cr_di,
  seed = 1,
  cond = 1,
)
}
}
