% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lr.R
\name{bootstrap_lr}
\alias{bootstrap_lr}
\title{Runs a parameteric bootstrapping likelihood ratio test on two DAISIE models}
\usage{
bootstrap_lr(
  daisie_data,
  data_name,
  model_1,
  model_2,
  array_index,
  cond,
  methode = "lsodes",
  optimmethod = "subplex",
  results_dir = NULL,
  test = FALSE
)
}
\arguments{
\item{daisie_data}{A list, conforming to the \code{\link{DAISIE}} object format.
Usually preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation
for details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_cr()}} and friends.}

\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{model_1}{A string with model that should run. For a list of options
see documentation for \code{model}  parameter in \code{\link[=run_daisie_ml]{run_daisie_ml()}}.}

\item{model_2}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_daisie_ml]{run_daisie_ml()}}.}

\item{array_index}{A single numeric with the array index. It is used for
naming the output file.}

\item{cond}{An integer specifying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \verb{odeint::}-prefix, \verb{\link{deSolve}{ode}()} method is
assumed. The default method overall is
\code{"lsodes"} for \verb{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \verb{\link[deSolve]\{ode\}()} for
\verb{\link{DAISIE_ML_IW}()}.}

\item{optimmethod}{Method used in likelihood optimization. Default is
\code{subplex} (see \verb{\link[subplex]\{subplex\}()} for full details).
Alternative is \code{"simplex"} which was the method in previous versions.}

\item{results_dir}{A string with the path to the directory where results
are to be stored or can be found. For example, if the data in question
is (to be) stored in \verb{folder_with_res/$data_name}, then \code{results_dir}
should be \code{"folder_with_res"}.
Defaults to \code{NULL}, which indicates the default directories are to be used.
Default directories are:
* \verb{$HOME/results/$data_name} if on the cluster
* \verb{getwd()/results/$data_name} if called from another environment
If \code{is.na(results_dir)}, then the object is returned to the R session and
not saved to file.}

\item{test}{A boolean, defaults to \code{FALSE}. Set to \code{TRUE} for testing
purposes, to fix the seed.}
}
\value{
Nothing. Writes bootstrapping results to a \code{.rds} file. This
file is stored in \code{file_path}. The directory in of \code{file_path} is created
if it doesn't exist.
}
\description{
Runs a parameteric bootstrapping likelihood ratio test on two DAISIE models
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
bootstrap_lr(
  daisie_data = Galapagos_datalist,
  model_1 = "cr_dd",
  model_2 = "cr_di",
  array_index = 1,
  cond = 1,
)
}
}
