% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bic.R
\name{calc_bic}
\alias{calc_bic}
\title{Compute Bayesian Information Criterion of a model}
\usage{
calc_bic(results, m)
}
\arguments{
\item{m}{Positive numeric integer with mainland species number.}
}
\value{
Numeric value with BIC
}
\description{
Compute Bayesian Information Criterion of a model
}
\author{
Joshua W. Lambert, Pedro Neves
}
