% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_loglik_ratio.R
\name{calc_loglik_ratio}
\alias{calc_loglik_ratio}
\title{Calculates the loglikelihood ratio between two models}
\usage{
calc_loglik_ratio(model_1_lik_res, model_2_lik_res)
}
\arguments{
\item{model_1_lik_res}{A data frame with results from a DAISIE maximum
likelihood model}

\item{model_2_lik_res}{A data frame with results from a DAISIE maximum
likelihood model}
}
\value{
Numeric likelihood ratio
}
\description{
Calculates the loglikelihood ratio between two models
}
