% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_value.R
\name{calc_p_value}
\alias{calc_p_value}
\title{Calculates p-value for rejecting a model from a distribution of likelihood
ratios}
\usage{
calc_p_value(data)
}
\arguments{
\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_cr()}} and friends.}
}
\value{
Numeric p-value
}
\description{
Calculates p-value for rejecting a model from a distribution of likelihood
ratios
}
