% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sim_metrics.R
\name{calc_sim_metrics}
\alias{calc_sim_metrics}
\title{Calculates summary metrics from a simulation}
\usage{
calc_sim_metrics(daisie_data)
}
\arguments{
\item{daisie_data}{A list, conforming to the \code{\link{DAISIE}} object format.
Usually preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation
for details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_cr()}} and friends.}
}
\value{
List of simulation metrics
}
\description{
Calculates summary metrics from a simulation
}
\author{
Joshua W. Lambert
}
