% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rep_seeds.R
\name{check_rep_seeds}
\alias{check_rep_seeds}
\title{Checks if jobs were run with the same seed}
\usage{
check_rep_seeds(logs_path)
}
\arguments{
\item{logs_path}{Character with path to the folder containing logs. Only log
files should be present, in plain text format.}
}
\value{
A data frame with four columns. Each line contains the information
of one result with a duplicated seed. There will be no lines if there are
no duplicated seeds in the any of the logs.
Columns are as follows:
\itemize{
\item \code{Data}: A character vector with the name of the data set where duplicates
were found.
\item \code{Models}: A numeric with corresponding array index. Will be empty if no
duplicates were found.
\item \code{Seeds}: A numeric with the corresponding seed that was duplicated.
\item \code{Array_indices}: A numeric with corresponding array index.
}
}
\description{
Checks every log file inside a folder for the record of the used seed.
Returns duplicated seeds and corresponding job arrays.
}
\note{
This function is the preferred method for checking for the presence of
repeated seeds.
However, it will fail if the log files were generated by older versions of
the package, as it expects that the seed and array information are always at
the same location. If you encounter issues, try running
\code{\link[=check_rep_seeds_depr]{check_rep_seeds_depr()}} instead. For all other cases, give preference
to this function as better optimization for log output parsing is possible.
}
\examples{
\dontrun{
repeated_seeds <- check_rep_seeds(logs_path = "/logs/")
}
}
\seealso{
\code{\link[=check_rep_seeds_depr]{check_rep_seeds_depr()}} for deprecated checking compatible
with older log files.
}
\author{
Pedro Neves
}
