% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repeated_seeds.R
\name{check_repeated_seeds}
\alias{check_repeated_seeds}
\title{Checks if jobs were run with the same seed}
\usage{
check_repeated_seeds(logs_path)
}
\arguments{
\item{logs_path}{Character with path to the folder containing logs. Only log
files should be present, in plain text format.}
}
\value{
A named list with two elements.
\itemize{
\item \code{duplicated_seeds}: A character vector with the seed values that were
found to be duplicated within the folder. Will be empty if no duplicates
were found
\item \code{duplicated_array_indices}: A numeric vector with corresponding array
indices. Will be empty if no duplicates were found.
}
}
\description{
Checks every log file inside a folder for the record of the used seed.
Returns duplicated seeds and corresponding job arrays.
}
\note{
This function is the preferred method for checking for the presence of
repeated seeds.
However, it will fail if the log files were generated by older versions of
the package, as it expects that the seed and array information are always at
the same location. If you encounter issues, try running
\code{\link[=check_repeated_seeds_long]{check_repeated_seeds_long()}} instead. For all other cases, give preference
to this function as better optimization for log output parsing is possible.
}
\examples{
\dontrun{
repeated_seeds <- check_repeated_seeds(logs_path = "/logs/")
}
}
\seealso{
\code{\link[=check_repeated_seeds_long]{check_repeated_seeds_long()}} for deprecated checking compatible
with older log files.
}
\author{
Pedro Neves
}
