% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_model.R
\name{choose_best_model}
\alias{choose_best_model}
\title{From multiple seeds, choose the best model fit}
\usage{
choose_best_model(model_lik_res)
}
\arguments{
\item{model_lik_res}{A data frame with results from a DAISIE maximum
likelihood model.}
}
\value{
A list of length one, with a data frame of 1 row containing the best
estimated models from the given seeds, as determined by the loglik.
If no model was estimated successfully, returns similar structure with all
fields as \code{NA}.
}
\description{
From multiple seeds, choose the best model fit
}
\author{
Pedro Santos Neves
}
