% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_rank.R
\name{col_rank}
\alias{col_rank}
\title{Stub}
\usage{
col_rank(data, sim, plot = FALSE)
}
\arguments{
\item{data}{stub}

\item{sim}{stub}

\item{plot}{stub}
}
\value{
Matrix?
}
\description{
Stub
}
\examples{
\dontrun{
col_rank(
data = Galapagos_datalist,
sim = islands_1type_1000reps)
}
}
