% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_dir.R
\name{create_results_dir_path}
\alias{create_results_dir_path}
\title{Create results directory path}
\usage{
create_results_dir_path(data_name, results_dir = NULL)
}
\arguments{
\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{results_dir}{A string with the path to the directory where results
are to be stored or can be found. For example, if the data in question
is (to be) stored in \verb{folder_with_res/$data_name}, then \code{results_dir}
should be \code{"folder_with_res"}.
Defaults to \code{NULL}, which indicates the default directories are to be used.
Default directories are:
* \verb{$HOME/results/$data_name} if on the cluster
* \verb{getwd()/results/$data_name} if called from another environment
If \code{is.na(results_dir)}, then the object is returned to the R session and
not saved to file.}
}
\value{
String with platform appropriate file path to be used for the results
directory for a given data set.
}
\description{
Create results directory path
}
\examples{
results_dir_path <- create_results_dir_path(data_name = "Azores")
}
\author{
Pedro Santos Neves
}
