% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIEutils-package.R
\docType{package}
\name{DAISIEutils-package}
\alias{DAISIEutils}
\alias{DAISIEutils-package}
\title{DAISIEutils: Utility Functions for the DAISIE Package}
\description{
The goal of DAISIEutils is to collect useful utility functions that are used recurently in DAISIE projects. DAISIEutils is a companion package of the R package DAISIE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tece-lab/DAISIEutils}
  \item \url{https://tece-lab.github.io/DAISIEutils/}
  \item Report bugs at \url{https://github.com/tece-lab/DAISIEutils/issues}
}

}
\author{
\strong{Maintainer}: Pedro Santos Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID})

Authors:
\itemize{
  \item Joshua W. Lambert \email{j.w.l.lambert@rug.nl} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID})
  \item Luis Valente \email{luis.valente@naturalis.nl} (\href{https://orcid.org/0000-0003-4247-8785}{ORCID})
  \item Richèl J.C. Bilderbeek \email{richel@richelbilderbeek.nl} (\href{https://orcid.org/0000-0003-1107-7049}{ORCID})
  \item Rampal S. Etienne \email{r.s.etienne@rug.nl} (\href{https://orcid.org/0000-0003-2142-7612}{ORCID})
}

}
\keyword{internal}
