% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameters documentation}
\usage{
default_params_doc(
  model,
  data_name,
  results_root_folder,
  data,
  seed,
  file_path,
  results,
  cond
)
}
\arguments{
\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0lama"} Clade specific model - diversity dependent and no
anagenesis (lama fixed to zero). All other parameters free.
\item \code{"cr_di_0lama"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (lama fixed to zero). All other parameters free.
\item \code{"cr_di_0lamc"} Clade specific model - diversity indendent (K = Inf)
and no cladogenesis (lamc fixed to zero). All other parameters free.
\item \code{"cr_dd_0lamc"} Clade specific model - diversity dependent and no
cladogenesis (lamc fixed to zero). All other parameters free.
\item \code{"rr_lamc_di"} Clade specific model - diversity independent (K = Inf)
and relaxed cladogenesis. All other parameters free.
\item \code{"rr_lamc_dd"} Clade specific model - diversity dependent and relaxed
cladogenesis. All other parameters free.
\item \code{"rr_mu_di"} Clade specific model - diversity independent (K = Inf) and
relaxed extinction. All other parameters free.
\item \code{"rr_mu_dd"} Clade specific model - diversity dependent and relaxed
extinction. All other parameters free.
\item \code{"rr_k"} Clade specific model - diversity dependent and relaxed carrying
capacity. All other parameters free.
\item \code{"rr_lama_di"} Clade specific model - diversity independent (K = Inf)
and relaxed anagenesis. All other parameters free.
\item \code{"rr_lama_dd"} Clade specific model - diversity dependent and relaxed
anagenesis. All other parameters free.
\item \code{"rr_mu_di_0lamc"} Clade specific model - diversity independent,
relaxed extinction, and no cladogenesis (lamc fixed to zero). All other
parameters free.
\item \code{"rr_mu_dd_0lamc"} Clade specific model - diversity dependent,
\item \code{"rr_k_0lamc"} Clade specific model - diversity dependent,
\item \code{"rr_lama_di_0lamc"} Clade specific model - diversity independent
\item \code{"r_lama_dd_0lamc"} Clade specific model - diversity dependent
\item \code{"rr_lamc_di_0lama"} Clade specific model - diversity independent
\item \code{"rr_lamc_dd_0lama"} Clade specific model - diversity dependent
\item \code{"rr_mu_di_0lama"} Clade specific model - diversity independent
\item \code{"rr_mu_dd_0lama"} Clade specific model - diversity dependent
\item \code{"rr_k_0lama"} Clade specific model - diversity dependent
}}

\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{results_root_folder}{Character. A path to the root folder containing
subfolders. Each subfolder contains result files for analysis runs.}

\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}} and friends.}

\item{seed}{A single numeric to be used as seed.}

\item{file_path}{The system directory where the output files will be stored.}

\item{results}{A data frame contaning model results as created by
\code{\link[=run_analysis]{run_analysis()}}. These results are the \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}} and the
bic, and they're saved as an RDS file by \code{\link[=run_analysis]{run_analysis()}}.}

\item{cond}{An integer specfying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}
}
\value{
Nothing
}
\description{
Default parameters documentation
}
\keyword{internal}
