% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_models.R
\name{get_available_models}
\alias{get_available_models}
\title{List all available models}
\usage{
get_available_models()
}
\value{
A character vector with the named codes for all available models.
}
\description{
List all available models
}
\examples{
available_models <- DAISIEutils:::get_available_models()
}
\author{
Pedro Neves
}
\keyword{internal}
