% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_fit.R
\name{goodness_fit}
\alias{goodness_fit}
\title{Calculates the metrics from simulated data and compares it to
empirical data}
\usage{
goodness_fit(data, sim, plot = FALSE)
}
\arguments{
\item{data}{stub}

\item{sim}{stub}

\item{plot}{stub}
}
\value{
List of numeric vectors
}
\description{
Calculates the metrics from simulated data and compares it to
empirical data
}
\examples{
\dontrun{
goodness_fit(
data = Galapagos_datalist,
sim = islands_1type_1000reps)
}
}
