% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_daisie_data.R
\name{is_daisie_data}
\alias{is_daisie_data}
\title{Checks if object is a \link{DAISIE} simulation data}
\usage{
is_daisie_data(daisie_data)
}
\value{
\code{TRUE} if \code{data} is a valid \link{DAISIE} object, \code{FALSE} if not.
}
\description{
Checks if object is a \link{DAISIE} simulation data
}
\author{
Pedro Neves, Richèl J. C. Bilderbeek
}
