% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_daisie_object.R
\name{is_daisie_object}
\alias{is_daisie_object}
\title{Checks if object is a \link{DAISIE} simulation object}
\usage{
is_daisie_object(data)
}
\value{
\code{TRUE} if \code{data} is a valid \link{DAISIE} object, \code{FALSE} if not.
}
\description{
Checks if object is a \link{DAISIE} simulation object
}
