% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_seed.R
\name{jump_seed}
\alias{jump_seed}
\title{Jumps forward in a seed stream}
\usage{
jump_seed(index, jump_size = 1e+08)
}
\arguments{
\item{index}{stub}

\item{jump_size}{stub}
}
\value{
Nothing, this is a void function.
}
\description{
Jumps forward in a seed stream
}
\examples{
\dontrun{
jump_seed(index = 1, jump_size = 1e8)
}
}
