% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_txt_results.R
\name{load_txt_results}
\alias{load_txt_results}
\title{Load analysis result in txt format}
\usage{
load_txt_results(results_root_folder)
}
\arguments{
\item{results_root_folder}{Character. A path to the root folder containing
subfolders. Each subfolder contains result files for analysis runs.}
}
\value{
Data frame with analysis results
}
\description{
Load analysis result in txt format
}
\author{
Pedro Neves
}
