% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_col_ranks.R
\name{plot_col_ranks}
\alias{plot_col_ranks}
\title{Plots the ranking of the colonisation events}
\usage{
plot_col_ranks(tt)
}
\arguments{
\item{tt}{stub}
}
\value{
Nothing, returns plot
}
\description{
Plots the ranking of the colonisation events
}
\examples{
\dontrun{
plot_col_ranks(tt = tt)
}
}
