% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ll_surface.R
\name{plot_ll_surface}
\alias{plot_ll_surface}
\title{Plot contour of loglikelihood surface around a given value}
\usage{
plot_ll_surface(data_to_plot, param_1, param_2)
}
\value{
A contour ggplot with the likelihood surface.
}
\description{
Plots the likelihood surface around a set of parameters \code{initparsopt} around
two axes (\code{param_1} and \code{param_2}).
}
\author{
Pedro Santos Neves
}
