% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_hist.R
\name{plot_metric_hist}
\alias{plot_metric_hist}
\title{Plots histograms for a given metric}
\usage{
plot_metric_hist(results, metric, type)
}
\arguments{
\item{results}{List output from \code{\link[=goodness_fit]{goodness_fit()}}}

\item{metric}{A string determining which metric to plot, options are:
\itemize{
\item num_spec
\item num_col
\item num_endemic_singletons
\item num_nonendemic
\item num_type_1_spec
\item num_type_2_spec
\item num_endemic_clades
\item largest_clade_size
\item num_col_incl_recol
\item num_recol
}}

\item{type}{String, eiter: "type1_only" or "type1and2"}
}
\value{
Nothing
}
\description{
Plots histograms for a given metric
}
\examples{
\dontrun{
goodness_fit(
metric = "num_spec",
type = "type1_only")
}
}
