% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_metrics.R
\name{plot_sim_metrics}
\alias{plot_sim_metrics}
\title{Plots metrics from simulations}
\usage{
plot_sim_metrics(sim_metrics)
}
\arguments{
\item{sim_metrics}{A list of metrics output from \code{\link[=calc_sim_metrics]{calc_sim_metrics()}}}
}
\value{
none Four plots are shown: histogram of the number of species,
histogram of the number of colonizations, histogram of the largest clade size
and histogram of the rank of the largest clade
}
\description{
Plots metrics from simulations
}
\author{
Joshua W. Lambert
}
