% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_main_header.R
\name{print_main_header}
\alias{print_main_header}
\title{Print session and run info to console/log file}
\usage{
print_main_header(run_name, m, metadata)
}
\arguments{
\item{run_name}{Character with name of run to be printed. Can be anything,
altough it is useful to name it after the dataset that is being analysed or
generated.}

\item{m}{Positive numeric integer with mainland species number.}

\item{metadata}{A numeric or a string. To be improved to take more info.}
}
\value{
Nothing. Prints session and run info used by DAISIE to console.
}
\description{
Useful to call as at the start of normal job scripts to  run on the cluster
so that metadata is recorded in the log files.
}
\note{
Message is used to print function arguments. To retain formatting,
a simple \code{sessioninfo::session_info()} is used for session info, which uses
print. Hence, \code{session_info()} will go on \code{stdout}, while the remaining
output of this function will go to \code{stderr}.
}
\examples{
\dontrun{
print_main_header(
  run_name = "example_name",
  M = 165,
  metadata = 1
)
}
}
\author{
Pedro Neves, Luis Valente
}
