% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_results.R
\name{read_model_results}
\alias{read_model_results}
\title{Reads \code{run_daisie_ml()} results}
\usage{
read_model_results(results_root_folder)
}
\arguments{
\item{results_root_folder}{Character. A path to the root folder containing
subfolders. Each subfolder contains result files for analysis runs.}
}
\value{
Data frame with all the results obtained
}
\description{
Reads \code{run_daisie_ml()} results
}
\author{
Pedro Neves
}
