% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seed.R
\name{read_seed}
\alias{read_seed}
\title{Read the seed from a saved file}
\usage{
read_seed()
}
\value{
Numeric vector with the state of the RNG
}
\description{
Read the seed from a saved file
}
\examples{
\dontrun{
read_seed()
}
}
\seealso{
Other seed: 
\code{\link{save_seed}()},
\code{\link{set_seed}()}
}
\concept{seed}
