% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_main.R
\name{run_main}
\alias{run_main}
\title{Run analysis from a set of initial parameters}
\usage{
run_main(datalist, m, model, seeds, verbose = FALSE)
}
\arguments{
\item{datalist}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details.}

\item{m}{Positive numeric integer with mainland species number.}

\item{model}{Numeric integer from 1 to 4 with model that should run. Models
are as follows:
\enumerate{
\item Clade specific model - diversity dependent. All parameters free.
\item Clade specific model - diversity independent (K = Inf). All other
parameters free.
\item Clade specific model - diversity dependent and no anagenesis (laa fixed
to zero). All other parameters free.
\item Clade specific model - diversity independent (K = Inf) and no
anagenesis (laa fixed to zero). All other parameters free.
}}

\item{seeds}{Numeric vector contaning the seeds to be used in the analysis.}

\item{verbose}{Boolean. Defaults to FALSE. If TRUE, session info and run.}
}
\value{
Saves results to .txt file.
}
\description{
Run analysis from a set of initial parameters
}
\examples{
# Run 1 set of randomly sampled initial parameters for scenario 1 and
# model 1 of the Galapagos datalist.
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
run_main(
  datalist = Galapagos_datalist,
  m = 50,
  model = 1,
  seeds = 1
)
}
}
\author{
Pedro Neves
}
