% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_seed.R
\name{save_seed}
\alias{save_seed}
\title{Saves the state of the RNG}
\usage{
save_seed(seed)
}
\arguments{
\item{seed}{The clock time set in \code{\link[=set_seed]{set_seed()}}.}
}
\value{
Nothing, only saves the state of the RNG in the global environment
}
\description{
Saves the state of the RNG
}
\examples{
\dontrun{
save_seed()
}
}
\seealso{
Other seed: 
\code{\link{read_seed}()},
\code{\link{set_seed}()}
}
\concept{seed}
