% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Read the results of \code{\link[=run_daisie_ml]{run_daisie_ml()}} and compares model selection}
\usage{
sensitivity(data_names, full_output = FALSE, save_to_file = TRUE)
}
\arguments{
\item{data_names}{A vector of strings with the names of the data sets you
want to compare sensitivity.}

\item{full_output}{A boolean determining whether the full model output is
returned.}

\item{save_to_file}{Boolean, defaults to \code{TRUE}. If \code{TRUE}, results will
saved to file. If \code{FALSE}, results will be returned to the environment.}
}
\value{
List
}
\description{
Reads the results of \code{\link[=run_daisie_ml]{run_daisie_ml()}} and compares model selection to
determine sensitivity to different data input into the same model or set of
models.
}
\examples{
\dontrun{
sensitivity(
  data_names = c("Azores", "Azores_alt_m"),
  full_output = FALSE
)
}
}
\author{
Joshua W. Lambert
}
