% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seed.R
\name{set_seed}
\alias{set_seed}
\title{Sets seed using current system time.}
\usage{
set_seed()
}
\value{
Nothing, system time and \code{.GlobalEnv$.Random.seed} to \code{.rds} file.
}
\description{
Sets seed using current system time.
}
\examples{
\dontrun{set_seed()}
}
\seealso{
Other seed: 
\code{\link{read_seed}()},
\code{\link{save_seed}()}
}
\concept{seed}
