% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{setup_model}
\alias{setup_model}
\title{Set up DAISIE_ML arguments}
\usage{
setup_model(model)
}
\arguments{
\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0lama"} Clade specific model - diversity dependent and no
anagenesis (lama fixed to zero). All other parameters free.
\item \code{"cr_di_0lama"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (lama fixed to zero). All other parameters free.
}}
}
\value{
A named list with \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}} arguments.
}
\description{
Set up DAISIE_ML arguments
}
\examples{
model <- "cr_dd"
setup_model(
  model = model
)
}
\author{
Luis M Valente, Pedro Neves, Joshua W. Lambert
}
