% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_standard_model_args.R
\name{setup_model_arguments}
\alias{setup_model_arguments}
\title{Set up DAISIE_ML arguments}
\usage{
setup_model_arguments(model, r_lamc, r_mu, r_k, r_gam, r_ana)
}
\value{
A list with \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}} arguments.
}
\description{
Set up DAISIE_ML arguments
}
\examples{
model <- 1
r_lamc <- 1
r_mu <- 2
r_k <- 3
r_gam <- 4
r_ana <- 5
setup_model_arguments(
  model = model,
  r_lamc = r_lamc,
  r_mu = r_mu,
  r_k = r_k,
  r_gam = r_gam,
  r_ana = r_ana
)
}
\author{
Luis M Valente, Pedro Neves
}
