% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_standard_model_args.R
\name{setup_standard_model_args}
\alias{setup_standard_model_args}
\title{Set up DAISIE_ML arguments}
\usage{
setup_standard_model_args(model, r_lamc, r_mu, r_k, r_gam, r_ana)
}
\arguments{
\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf) and
no anagenesis (laa fixed to zero). All other parameters free.
}}

\item{r_lamc}{Numeric. Uniformly sampled value of cladogenesis rate.}

\item{r_mu}{Numeric. Uniformly sampled value of extinction rate.}

\item{r_k}{Numeric. Uniformly sampled value of carrying capacity.}

\item{r_gam}{Numeric. Uniformly sampled value of immigration rate.}

\item{r_ana}{Numeric. Uniformly sampled value of anagenesis rate.}
}
\value{
A named list with \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}} arguments.
}
\description{
Set up DAISIE_ML arguments
}
\examples{
model <- "cr_dd"
r_lamc <- 1
r_mu <- 2
r_k <- 3
r_gam <- 4
r_ana <- 5
setup_standard_model_args(
  model = model,
  r_lamc = r_lamc,
  r_mu = r_mu,
  r_k = r_k,
  r_gam = r_gam,
  r_ana = r_ana
)
}
\author{
Luis M Valente, Pedro Neves, Joshua W. Lambert
}
