% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_goodness_of_fit.R
\name{summarize_bootstrap_results}
\alias{summarize_bootstrap_results}
\title{Computes summary results of bootstrap simulations useful for plotting}
\usage{
summarize_bootstrap_results(
  simulation_dataset,
  empirical_summary_statistics,
  m_pool = 1000
)
}
\arguments{
\item{simulation_dataset}{output of DAISIE_sim}

\item{empirical_summary_statistics}{vector with the number of species, number
of colonization, size of the largest clade and the rank of the largest clade
in the empirical data}

\item{m_pool}{size of the mainland pool}

\item{time}{age of the island or archipelago}
}
\value{
overall_results a list of results
}
\description{
The output is a list of results
}
\examples{
\dontrun{
dataset_CS <- 'd:/data/tex/Etienne et al 2019/frogs_sim_CS_T30_M1000_R5000.RData'
dataset_IW <- 'd:/data/tex/Etienne et al 2019/frogs_sim_IW_T30_M1000_R5000.RData'
overall_results_CS <- DAISIEutils::summarize_bootstrap_results(
  simulation_dataset = dataset_CS
)
overall_results_IW <- DAISIEutils::summarize_bootstrap_results(
  simulation_dataset = dataset_IW
)
par(mfrow = c(2, 4), cex.lab = 1.5, cex.main = 1.5)
DAISIEutils::plot_bootstrap_results(
  overall_results = overall_results_CS,
  title = 'Simulated under CS'
)
DAISIEutils::plot_bootstrap_results(
  overall_results = overall_results_IW,
  title = 'Simulated under IW'
)
}
}
\author{
Rampal S. Etienne & Luis Valente
}
