% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lr.R
\name{bootstrap_lr}
\alias{bootstrap_lr}
\title{Runs a parameteric bootstrapping likelihood ratio test on two DAISIE models}
\usage{
bootstrap_lr(
  data,
  data_name,
  model_1,
  model_2,
  array_index,
  cond,
  test = FALSE
)
}
\arguments{
\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}} and friends.}

\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{model_1}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_daisie_ml]{run_daisie_ml()}}.}

\item{model_2}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_daisie_ml]{run_daisie_ml()}}.}

\item{array_index}{A single numeric with the array index. It is used for
naming the output file.}

\item{cond}{An integer specfying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}

\item{test}{A boolean, defaults to \code{FALSE}. Set to \code{TRUE} for testing
purposes, to fix the seed.}
}
\value{
Nothing. Writes bootstrapping results to a \code{.rds} file. This
file is stored in \code{file_path}. The directory in of \code{file_path} is created
if it doesn't exist.
}
\description{
Runs a parameteric bootstrapping likelihood ratio test on two DAISIE models
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
bootstrap_lr(
  data = Galapagos_datalist,
  model_1 = "cr_dd",
  model_2 = "cr_di",
  array_index = 1,
  cond = 1,
)
}
}
