% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bic.R
\name{calc_bic}
\alias{calc_bic}
\title{Compute Bayesian Information Criterion of a model}
\usage{
calc_bic(results, m)
}
\arguments{
\item{results}{A tibble contaning model results as read by
\code{\link[=load_txt_results]{load_txt_results()}}. These results are themselves created and saved as
text files by \code{\link[=run_analysis]{run_analysis()}} and friends.}
}
\value{
Numeric value with BIC
}
\description{
Compute Bayesian Information Criterion of a model
}
\author{
Joshua W. Lambert, Pedro Neves
}
