% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bic.R
\name{calc_bic}
\alias{calc_bic}
\title{Compute Bayesian Information Criterion of a model}
\usage{
calc_bic(results, data)
}
\arguments{
\item{results}{A data frame containing model results as created by
\code{\link[=run_daisie_ml]{run_daisie_ml()}}. These results are the \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}} and the
bic, and they're saved as an RDS file by \code{\link[=run_daisie_ml]{run_daisie_ml()}}.}

\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}} and friends.}
}
\value{
Numeric value with BIC
}
\description{
Compute Bayesian Information Criterion of a model
}
\author{
Joshua W. Lambert, Pedro Neves
}
