% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{calc_power}
\alias{calc_power}
\title{Calculates power to detect the true model}
\usage{
calc_power(daisie_data, results_dir = NULL)
}
\arguments{
\item{daisie_data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_cr()}} and friends.}

\item{results_dir}{A string with the path to the directory where results
are to be stored or can be found. For example, if the data in question
is (to be) stored in \verb{folder_with_res/$data_name}, then \code{results_dir}
should be \code{"folder_with_res"}.
Defaults to \code{NULL}, which indicates the default directories are to be used.
Default directories are:
* \verb{$HOME/results/$data_name} if on the cluster
* \verb{getwd()/results/$data_name} if called from another environment}
}
\value{
Numeric power
}
\description{
Calculates power to detect the true model
}
