% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sim_metrics.R
\name{calc_sim_metrics}
\alias{calc_sim_metrics}
\title{Calculates summary metrics from a simulation}
\usage{
calc_sim_metrics(daisie_data)
}
\value{
List of simulation metrics
}
\description{
Calculates summary metrics from a simulation
}
\author{
Joshua W. Lambert
}
