% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repeated_seeds_long.R
\name{check_repeated_seeds_long}
\alias{check_repeated_seeds_long}
\title{Checks if jobs were run with the same seed on older logs}
\usage{
check_repeated_seeds_long(logs_path)
}
\arguments{
\item{logs_path}{Character with path to the folder containing logs. Only log
files should be present, in plain text format.}
}
\value{
A data frame with four columns. Each line contains the information
of one result with a duplicated seed. There will be no lines if there are
no duplicated seeds in the any of the logs.
Columns are as follows:
\itemize{
\item \code{Data}: A character vector with the name of the data set where duplicates
were found.
\item \code{Models}: A numeric with corresponding array index. Will be empty if no
duplicates were found.
\item \code{Seeds}: A numeric with the corresponding seed that was duplicated.
\item \code{Array_indices}: A numeric with corresponding array index.
}
}
\description{
Checks every log file inside a folder for the record of the used seed and
returns duplicated seeds and corresponding job arrays.
}
\note{
In this function, the lines containing the scraped values can be anywhere in
each file, as this occurred in older versions of the package. Checking is
thus less efficient and preference should be given to
\code{\link[=check_repeated_seeds]{check_repeated_seeds()}} in most cases, unless older log files are being
checked. \code{\link[=check_repeated_seeds]{check_repeated_seeds()}} will fail for most older log files.
}
\examples{
\dontrun{
repeated_seeds <- check_repeated_seeds_long(logs_path = "/logs/")
}
}
\seealso{
\code{\link[=check_repeated_seeds]{check_repeated_seeds()}} for optimized checks on current logs.
}
\author{
Pedro Neves
}
