% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_folder.R
\name{create_output_folder}
\alias{create_output_folder}
\title{Create model output folder}
\usage{
create_output_folder(data_name, model, seed)
}
\arguments{
\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf) and
no anagenesis (laa fixed to zero). All other parameters free.
}}

\item{seed}{A single numeric to be used as seed.}
}
\value{
Creates appropriate directory. Returns string with path for
output object. If on Peregrine, folder will be \verb{$HOME/results/$output_name}.
If called from another environment, folder will be
\verb{getwd()/results/$output_name}.
}
\description{
Create model output folder
}
\examples{
\dontrun{
create_output_folder(output_name = "results_folder", model = 1)
}
}
\author{
Pedro Neves
}
