% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_folder.R
\name{create_output_folder}
\alias{create_output_folder}
\title{Create output folder}
\usage{
create_output_folder(data_name, results_dir = NULL)
}
\arguments{
\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{results_dir}{A string with the path to the directory where results
are to be stored or can be found. For example, if the data in question
is (to be) stored in \verb{folder_with_res/$data_name}, then \code{results_dir}
should be \code{"folder_with_res"}.
Defaults to \code{NULL}, which indicates the default directories are to be used.
Default directories are:
* \verb{$HOME/results/$data_name} if on the cluster
* \verb{getwd()/results/$data_name} if called from another environment
If \code{is.na(results_dir)}, then the object is returned to the R session and
not saved to file.}
}
\value{
Creates appropriate directory. Returns string with path for output object. By
default, if on Hábrók, folder will be \verb{$HOME/results/$data_name}. If
called from another environment, folder will be
\verb{getwd()/results/$data_name}. Alternatively, another valid root can be
specified, resulting in \verb{results_dir/$data_name}.
}
\description{
Create output folder
}
\examples{
\dontrun{
create_output_folder(
  data_name = "results_folder"
)
}
}
\author{
Pedro Santos Neves
}
