% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_folder.R
\name{create_output_folder}
\alias{create_output_folder}
\title{Create model output folder}
\usage{
create_output_folder(output_name, model, verbose = FALSE)
}
\arguments{
\item{model}{Numeric integer from 1 to 4 with model that should run. Models
are as follows:
\enumerate{
\item Clade specific model - diversity dependent. All parameters free.
\item Clade specific model - diversity independent (K = Inf). All other
parameters free.
\item Clade specific model - diversity dependent and no anagenesis (laa fixed
to zero). All other parameters free.
\item Clade specific model - diversity independent (K = Inf) and no
anagenesis (laa fixed to zero). All other parameters free.
}}
}
\value{
Creates appropriate directory. Returns string with path for
output object. If on Peregrine, folder will be \verb{$HOME/results/$output_name}.
If called from another environment, folder will be
\verb{getwd()/results/$output_name}.
}
\description{
Create model output folder
}
\examples{
\dontrun{
create_output_folder(output_name = "results_folder", model = 1)
}
}
\author{
Pedro Neves
}
