% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIEutils-package.R
\docType{package}
\name{DAISIEutils-package}
\alias{DAISIEutils}
\alias{DAISIEutils-package}
\title{DAISIEutils: Utility Functions for the DAISIE Package}
\description{
The goal of DAISIEutils is to collect useful utility functions that
    are used reccurently in DAISIE projects. DAISIEutils is a companion package 
    of the R package DAISIE.
}
\author{
\strong{Maintainer}: Pedro Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID})

Authors:
\itemize{
  \item Joshua W. Lambert \email{j.w.l.lambert@rug.nl} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID})
  \item Luis Valente \email{luis.valente@naturalis.nl} (\href{https://orcid.org/0000-0003-4247-8785}{ORCID})
  \item Richèl J.C. Bilderbeek \email{richel@richelbilderbeek.nl} (\href{https://orcid.org/0000-0003-1107-7049}{ORCID})
}

}
\keyword{internal}
