% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameters documentation}
\usage{
default_params_doc(
  run_name,
  m,
  metadata,
  model,
  r_lamc,
  r_mu,
  r_k,
  r_gam,
  r_ana,
  output_name,
  results_root_folder,
  datalist,
  datalist_name,
  seed,
  seeds,
  file_path,
  results
)
}
\arguments{
\item{run_name}{Character with name of run to be printed. Can be anything,
altough it is useful to name it after the dataset that is being analysed or
generated.}

\item{m}{Positive numeric integer with mainland species number.}

\item{metadata}{A numeric or a string. To be improved to take more info.}

\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf) and
no anagenesis (laa fixed to zero). All other parameters free.
}}

\item{r_lamc}{Numeric. Uniformly sampled value of cladogenesis rate.}

\item{r_mu}{Numeric. Uniformly sampled value of extinction rate.}

\item{r_k}{Numeric. Uniformly sampled value of carrying capacity.}

\item{r_gam}{Numeric. Uniformly sampled value of immigration rate.}

\item{r_ana}{Numeric. Uniformly sampled value of anagenesis rate.}

\item{output_name}{A string. Will be the name of the created output folder.}

\item{results_root_folder}{Character. A path to the root folder containing
subfolders. Each subfolder contains result files for analysis runs.}

\item{datalist}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details.}

\item{datalist_name}{A character vector with the name of the datalist that
is the analyses run on. It is required internally so that it can be stored
in the first column of the output file.}

\item{seed}{A single numeric to be used as seed.}

\item{seeds}{Numeric vector contaning the seeds to be used in the analysis.}

\item{file_path}{The system directory where the output files will be stored.}

\item{results}{A tibble contaning model results as read by
\code{\link[=load_txt_results]{load_txt_results()}}. These results are themselves created and saved as
text files by \code{\link[=run_analysis]{run_analysis()}} and friends.}
}
\value{
Nothing
}
\description{
Default parameters documentation
}
\keyword{internal}
