% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameters documentation}
\usage{
default_params_doc(
  model,
  data_name,
  results_root_folder,
  data,
  seed,
  file_path,
  results,
  cond,
  model_1,
  model_2,
  model_1_lik_res,
  model_2_lik_res,
  model_lik_res,
  lik_res,
  data_names,
  full_output
)
}
\arguments{
\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0lac"} Clade specific model - diversity indendent (K = Inf)
and no cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"cr_dd_0lac"} Clade specific model - diversity dependent and no
cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"rr_lac_di"} Clade specific model - diversity independent (K = Inf)
and relaxed cladogenesis. All other parameters free.
\item \code{"rr_lac_dd"} Clade specific model - diversity dependent and relaxed
cladogenesis. All other parameters free.
\item \code{"rr_mu_di"} Clade specific model - diversity independent (K = Inf) and
relaxed extinction. All other parameters free.
\item \code{"rr_mu_dd"} Clade specific model - diversity dependent and relaxed
extinction. All other parameters free.
\item \code{"rr_k"} Clade specific model - diversity dependent and relaxed carrying
capacity. All other parameters free.
\item \code{"rr_laa_di"} Clade specific model - diversity independent (K = Inf)
and relaxed anagenesis. All other parameters free.
\item \code{"rr_laa_dd"} Clade specific model - diversity dependent and relaxed
anagenesis. All other parameters free.
\item \code{"rr_mu_di_0lac"} Clade specific model - diversity independent,
relaxed extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_mu_dd_0lac"} Clade specific model - diversity dependent, relaxed
extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_k_0lac"} Clade specific model - diversity dependent, relaxed
carrying capacity, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_laa_di_0lac"} Clade specific model - diversity independent,
relaxed anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"r_laa_dd_0lac"} Clade specific model - diversity dependent, relaxed
anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_lac_di_0laa"} Clade specific model - diversity independent,
relaxed cladogenesis, and no anagenesis (laa fixed to zero)
\item \code{"rr_lac_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_mu_di_0laa"} Clade specific model - diversity independent
\item \code{"rr_mu_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_k_0laa"} Clade specific model - diversity dependent
}}

\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{results_root_folder}{Character. A path to the root folder containing
subfolders. Each subfolder contains result files for analysis runs.}

\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}} and friends.}

\item{seed}{A single numeric to be used as seed.}

\item{file_path}{The system directory where the output files will be stored.}

\item{results}{A data frame contaning model results as created by
\code{\link[=run_analysis]{run_analysis()}}. These results are the \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}} and the
bic, and they're saved as an RDS file by \code{\link[=run_analysis]{run_analysis()}}.}

\item{cond}{An integer specfying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}

\item{model_1}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_analysis]{run_analysis()}}.}

\item{model_2}{A string with model that should run. For a list of options
see documentation for \code{model} parameter in \code{\link[=run_analysis]{run_analysis()}}.}

\item{model_1_lik_res}{A data frame with results from a DAISIE maximum
likelihood model}

\item{model_2_lik_res}{A data frame with results from a DAISIE maximum
likelihood model}

\item{model_lik_res}{A data frame with results from a DAISIE maximum
likelihood model}

\item{lik_res}{A data frame with results from a DAISIE maximum likelihood
model}

\item{data_names}{A vector of strings with the names of the data sets you
want to compare sensitivity}

\item{full_output}{A boolean determining whether the full model ouput is
returned in \code{\link[=sensitivity]{sensitivity()}}.}
}
\value{
Nothing
}
\description{
Default parameters documentation
}
\keyword{internal}
