% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_goodness_of_fit.R
\name{plot_bootstrap_results}
\alias{plot_bootstrap_results}
\title{Computes summary results of bootstrap simulations useful for plotting}
\usage{
plot_bootstrap_results(
  overall_results,
  sumstats = c(65, 5, 28, 1),
  ylim4 = 0.7,
  title = NULL
)
}
\arguments{
\item{overall_results}{summary results obtained with summarize_boorstrap_results}

\item{sumstats}{vector with the number of species, number
of colonization, size of the largest clade and the rank of the largest clade
in the empirical data}

\item{ylim4}{maximum of the plot of the rank of the largest clade}

\item{title}{title of the plot}
}
\value{
none Four plots are shown: histogram of the number of species,
histogram of the number of colonizations, histogram of the largest clade size
and histogram of the rank of the largest clade
}
\description{
The output is a list of results
}
\author{
Rampal S. Etienne & Luis Valente
}
