% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_metadata.R
\name{print_metadata}
\alias{print_metadata}
\title{Print session and run info to console/log file}
\usage{
print_metadata(data_name, model, seed)
}
\arguments{
\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0lama"} Clade specific model - diversity dependent and no
anagenesis (lama fixed to zero). All other parameters free.
\item \code{"cr_di_0lama"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (lama fixed to zero). All other parameters free.
}}

\item{seed}{A single numeric to be used as seed.}
}
\value{
Nothing. Prints session and run info used by DAISIE to console.
}
\description{
Useful to call as at the start of normal job scripts to  run on the cluster
so that metadata is recorded in the log files.
}
\note{
Message is used to print function arguments. To retain formatting,
a simple \code{sessioninfo::session_info()} is used for session info, which uses
print. Hence, \code{session_info()} will go on \code{stdout}, while the remaining
output of this function will go to \code{stderr}.
}
\examples{
\dontrun{
print_metadata(
  data_name = "Galapagos_datalist",
  model = "cr_di",
  seed = 1
)
}
}
\author{
Pedro Neves, Luis Valente, Joshua W. Lambert
}
