% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_metadata.R
\name{print_metadata}
\alias{print_metadata}
\title{Print session and run info to console/log file}
\usage{
print_metadata(data_name, model, array_index, seed, methode, optimmethod)
}
\arguments{
\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0lac"} Clade specific model - diversity indendent (K = Inf)
and no cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"cr_dd_0lac"} Clade specific model - diversity dependent and no
cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"rr_lac_di"} Clade specific model - diversity independent (K = Inf)
and relaxed cladogenesis. All other parameters free.
\item \code{"rr_lac_dd"} Clade specific model - diversity dependent and relaxed
cladogenesis. All other parameters free.
\item \code{"rr_mu_di"} Clade specific model - diversity independent (K = Inf) and
relaxed extinction. All other parameters free.
\item \code{"rr_mu_dd"} Clade specific model - diversity dependent and relaxed
extinction. All other parameters free.
\item \code{"rr_k"} Clade specific model - diversity dependent and relaxed carrying
capacity. All other parameters free.
\item \code{"rr_laa_di"} Clade specific model - diversity independent (K = Inf)
and relaxed anagenesis. All other parameters free.
\item \code{"rr_laa_dd"} Clade specific model - diversity dependent and relaxed
anagenesis. All other parameters free.
\item \code{"rr_mu_di_0lac"} Clade specific model - diversity independent,
relaxed extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_mu_dd_0lac"} Clade specific model - diversity dependent, relaxed
extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_k_0lac"} Clade specific model - diversity dependent, relaxed
carrying capacity, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_laa_di_0lac"} Clade specific model - diversity independent,
relaxed anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"r_laa_dd_0lac"} Clade specific model - diversity dependent, relaxed
anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_lac_di_0laa"} Clade specific model - diversity independent,
relaxed cladogenesis, and no anagenesis (laa fixed to zero)
\item \code{"rr_lac_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_mu_di_0laa"} Clade specific model - diversity independent
\item \code{"rr_mu_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_k_0laa"} Clade specific model - diversity dependent
}}

\item{array_index}{A single numeric with the array index. It is used for
naming the output file.}

\item{seed}{Integer with value to be used as the seed for Mersenne-Twister.
This value is determined by \code{\link[=Sys.time]{Sys.time()}} and \code{array_index} to ensure
parallel jobs have different seeds. Only the last 6 digits of the
\code{\link[=Sys.time]{Sys.time()}} (as an integer) are used.}

\item{methode}{Method for numerical integration of ODE system. Default is
\code{"odeint::runge_kutta_fehlberg78"}. See help for \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}
for a full list of available options.}

\item{optimmethod}{Method used in likelihood optimization. Default is
\code{"subplex"} (from \link{subplex-package}). Alternative is \code{simplex} which was the
method in previous versions.}
}
\value{
Nothing. Prints session and run info used by DAISIE to console.
}
\description{
Useful to call as at the start of normal job scripts to  run on the cluster
so that metadata is recorded in the log files.
}
\note{
Message is used to print function arguments. To retain formatting,
a simple \code{sessioninfo::session_info()} is used for session info, which uses
print. Hence, \code{session_info()} will go on \code{stdout}, while the remaining
output of this function will go to \code{stderr}.
}
\examples{
\dontrun{
print_metadata(
  data_name = "Galapagos_datalist",
  model = "cr_di",
  array_index = 1,
  seed = 1
)
}
}
\author{
Pedro Santos Neves, Luis Valente, Joshua W. Lambert
}
