% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_results.R
\name{read_model_results}
\alias{read_model_results}
\title{Reads \code{run_daisie_ml()} results}
\usage{
read_model_results(results_root_folder)
}
\arguments{
\item{results_root_folder}{Character. A path to the root folder containing
subfolders. Each subfolder contains result files for analysis runs.}
}
\value{
Nested list. First layer corresponds to the data sets, as per the
folders found in \code{results_root_folder}. The second layer corresponds to the
models run for each dataset, each containing a 1 row long data frame with
the of the alternate seed runs for each model.
}
\description{
Reads \code{run_daisie_ml()} results
}
\author{
Pedro Neves
}
