% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run spider analysis}
\usage{
run_analysis(datalist, m, model, seed, file_path)
}
\arguments{
\item{m}{Positive numeric integer with mainland species number.}

\item{model}{Numeric integer from 1 to 4 with model that should run. Models
are as follows:
\enumerate{
\item Clade specific model - diversity dependent. All parameters free.
\item Clade specific model - diversity independent (K = Inf). All other
parameters free.
\item Clade specific model - diversity dependent and no anagenesis (laa fixed
to zero). All other parameters free.
\item Clade specific model - diversity independent (K = Inf) and no
anagenesis (laa fixed to zero). All other parameters free.
}}
}
\value{
Nothing. Writes to file.
}
\description{
Run spider analysis
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
run_analysis(
  datalist = Galapagos_datalist,
  m = 165,
  model = 1,
  seed = 1,
  file_path
)
}
}
