% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run DAISIE analysis}
\usage{
run_analysis(data, model, seed)
}
\arguments{
\item{model}{Numeric integer from 1 to 4 with model that should run. Models
are as follows:
\enumerate{
\item Clade specific model - diversity dependent. All parameters free.
\item Clade specific model - diversity independent (K = Inf). All other
parameters free.
\item Clade specific model - diversity dependent and no anagenesis (laa fixed
to zero). All other parameters free.
\item Clade specific model - diversity independent (K = Inf) and no
anagenesis (laa fixed to zero). All other parameters free.
}}

\item{seed}{A single numeric to be used as seed.}
}
\value{
Nothing. Writes \code{\link{DAISIE}} analysis results to a \code{.txt} file. This
file is stored in \code{file_path}. The directory in of \code{file_path} is created
if it doesn't exist.
}
\description{
Run DAISIE analysis
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
run_analysis(
  datalist = Galapagos_datalist,
  m = 165,
  model = 1,
  seed = 1,
  file_path
)
}
}
\author{
Pedro Neves, Luis Valente
}
