% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run DAISIE analysis}
\usage{
run_analysis(data, model, seed, cond)
}
\arguments{
\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_constant_rate()}} and friends.}

\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf) and
no anagenesis (laa fixed to zero). All other parameters free.
}}

\item{seed}{A single numeric to be used as seed.}

\item{cond}{An integer specfying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}
}
\value{
Nothing. Writes \code{\link{DAISIE}} analysis results to a \code{.txt} file. This
file is stored in \code{file_path}. The directory in of \code{file_path} is created
if it doesn't exist.
}
\description{
Run DAISIE analysis
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
run_analysis(
  datalist = Galapagos_datalist,
  m = 165,
  model = cr_dd,
  seed = 1,
  file_path
)
}
}
\author{
Pedro Neves, Joshua W. Lambert, Luis Valente
}
