% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_daisie_ml.R
\name{run_daisie_ml}
\alias{run_daisie_ml}
\title{Run DAISIE analysis}
\usage{
run_daisie_ml(
  data,
  data_name,
  model,
  array_index,
  cond,
  methode = "odeint::runge_kutta_fehlberg78",
  optimmethod = "subplex",
  test = FALSE
)
}
\arguments{
\item{data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_cr()}} and friends.}

\item{data_name}{String. Will be used for the name of the created output
folder.}

\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0lac"} Clade specific model - diversity indendent (K = Inf)
and no cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"cr_dd_0lac"} Clade specific model - diversity dependent and no
cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"rr_lac_di"} Clade specific model - diversity independent (K = Inf)
and relaxed cladogenesis. All other parameters free.
\item \code{"rr_lac_dd"} Clade specific model - diversity dependent and relaxed
cladogenesis. All other parameters free.
\item \code{"rr_mu_di"} Clade specific model - diversity independent (K = Inf) and
relaxed extinction. All other parameters free.
\item \code{"rr_mu_dd"} Clade specific model - diversity dependent and relaxed
extinction. All other parameters free.
\item \code{"rr_k"} Clade specific model - diversity dependent and relaxed carrying
capacity. All other parameters free.
\item \code{"rr_laa_di"} Clade specific model - diversity independent (K = Inf)
and relaxed anagenesis. All other parameters free.
\item \code{"rr_laa_dd"} Clade specific model - diversity dependent and relaxed
anagenesis. All other parameters free.
\item \code{"rr_mu_di_0lac"} Clade specific model - diversity independent,
relaxed extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_mu_dd_0lac"} Clade specific model - diversity dependent, relaxed
extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_k_0lac"} Clade specific model - diversity dependent, relaxed
carrying capacity, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_laa_di_0lac"} Clade specific model - diversity independent,
relaxed anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"r_laa_dd_0lac"} Clade specific model - diversity dependent, relaxed
anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_lac_di_0laa"} Clade specific model - diversity independent,
relaxed cladogenesis, and no anagenesis (laa fixed to zero)
\item \code{"rr_lac_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_mu_di_0laa"} Clade specific model - diversity independent
\item \code{"rr_mu_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_k_0laa"} Clade specific model - diversity dependent
}}

\item{array_index}{A single numeric with the array index. It is used for
naming the output file.}

\item{cond}{An integer specifying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}

\item{optimmethod}{Method used in likelihood optimization. Default is
\code{"subplex"} (from \link{subplex-package}). Alternative is \code{simplex} which was the
method in previous versions.}

\item{test}{A boolean, defaults to \code{FALSE}. Set to \code{TRUE} for testing
purposes, to fix the seed.}
}
\value{
Nothing. Writes \code{\link{DAISIE}} analysis results to a \code{.rds} file. This
file is stored in \code{file_path}. The directory in of \code{file_path} is created
if it doesn't exist.
}
\description{
Run DAISIE analysis
}
\examples{
\dontrun{
data(Galapagos_datalist, package = "DAISIE")
run_daisie_ml(
  data = Galapagos_datalist,
  data_name = "Galapagos_datalist",
  model = "cr_dd",
  array_index = 1,
  cond = 1
)
}
}
\author{
Pedro Santos Neves, Joshua W. Lambert, Luis Valente
}
