% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_seed.R
\name{save_seed}
\alias{save_seed}
\title{Saves the state of the RNG}
\usage{
save_seed()
}
\value{
Nothing, only saves the state of the RNG in the global environment
}
\description{
Saves the state of the RNG
}
\examples{
\dontrun{
save_seed()
}
}
