% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Read the results of \code{\link[=run_daisie_ml]{run_daisie_ml()}} and compares model selection}
\usage{
sensitivity(data_names, full_output = FALSE, results_dir = NULL)
}
\arguments{
\item{data_names}{A vector of strings with the names of the data sets you
want to compare sensitivity.}

\item{full_output}{A boolean determining whether the full model output is
returned.}

\item{results_dir}{A string with the path to the directory where results
are to be stored or can be found. For example, if the data in question
is (to be) stored in \verb{folder_with_res/$data_name}, then \code{results_dir}
should be \code{"folder_with_res"}.
Defaults to \code{NULL}, which indicates the default directories are to be used.
Default directories are:
* \verb{$HOME/results/$data_name} if on the cluster
* \verb{getwd()/results/$data_name} if called from another environment
If \code{is.na(results_dir)}, then the object is returned to the R session and
not saved to file.}
}
\value{
A list of 3 elements if \code{full_output} is \code{FALSE}, or of 4 elements if
\code{TRUE}. The elements are as follows:
\describe{
\item{\code{best_fit_sensitivity}}{A character vector of length one, which
reports whether the best fit model is sensitive to the input.}
\item{\code{model_selection_sensitivity}}{A character vector of length one,
which reports whether the rank (or order) of the model selection is
sensitive to the input.}
\item{\code{model_selection_rank}}{A named list of as many elements as the there
are models in \code{data_names}. Each named list contains a sorted named vector
with the corresponding BIC value of each fit model. The sort is always
ascending.}
\item{\code{full_output}}{Only returned if \code{full_output} is \code{TRUE}. A named list
with a similar structure as \code{model_selection_rank}. Instead of named
vectors with the BIC values, however, the full \code{\link[=run_daisie_ml]{run_daisie_ml()}} data frame
output is returned. This is a one row data frame, with the parameter
estimates, degrees of freedom, convergence information and BIC value.}
}
}
\description{
Reads the results of \code{\link[=run_daisie_ml]{run_daisie_ml()}} and compares model selection to
determine sensitivity to different data input into the same model or set of
models. The \code{\link[=run_daisie_ml]{run_daisie_ml()}} results are expected to be located
in sub folders within the \verb{results/} folder in the current working directory.
These sub folders must have the same names as the elements of \code{data_names}.
}
\examples{
\dontrun{
sensitivity(
  data_names = c("Azores", "Azores_alt_m"),
  full_output = FALSE
)
}
}
\author{
Joshua W. Lambert
}
