% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{setup_model}
\alias{setup_model}
\title{Set up DAISIE_ML arguments}
\usage{
setup_model(model)
}
\arguments{
\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0lama"} Clade specific model - diversity dependent and no
anagenesis (lama fixed to zero). All other parameters free.
\item \code{"cr_di_0lama"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (lama fixed to zero). All other parameters free.
\item \code{"cr_di_0lamc"} Clade specific model - diversity indendent (K = Inf)
and no cladogenesis (lamc fixed to zero). All other parameters free.
\item \code{"cr_dd_0lamc"} Clade specific model - diversity dependent and no
cladogenesis (lamc fixed to zero). All other parameters free.
\item \code{"rr_lamc_di"} Clade specific model - diversity independent (K = Inf)
and relaxed cladogenesis. All other parameters free.
\item \code{"rr_lamc_dd"} Clade specific model - diversity dependent and relaxed
cladogenesis. All other parameters free.
\item \code{"rr_mu_di"} Clade specific model - diversity independent (K = Inf) and
relaxed extinction. All other parameters free.
\item \code{"rr_mu_dd"} Clade specific model - diversity dependent and relaxed
extinction. All other parameters free.
\item \code{"rr_k"} Clade specific model - diversity dependent and relaxed carrying
capacity. All other parameters free.
\item \code{"rr_lama_di"} Clade specific model - diversity independent (K = Inf)
and relaxed anagenesis. All other parameters free.
\item \code{"rr_lama_dd"} Clade specific model - diversity dependent and relaxed
anagenesis. All other parameters free.
\item \code{"rr_mu_di_0lamc"} Clade specific model - diversity independent,
relaxed extinction, and no cladogenesis (lamc fixed to zero). All other
parameters free.
\item \code{"rr_mu_dd_0lamc"} Clade specific model - diversity dependent,
\item \code{"rr_k_0lamc"} Clade specific model - diversity dependent,
\item \code{"rr_lama_di_0lamc"} Clade specific model - diversity independent
\item \code{"r_lama_dd_0lamc"} Clade specific model - diversity dependent
\item \code{"rr_lamc_di_0lama"} Clade specific model - diversity independent
\item \code{"rr_lamc_dd_0lama"} Clade specific model - diversity dependent
\item \code{"rr_mu_di_0lama"} Clade specific model - diversity independent
\item \code{"rr_mu_dd_0lama"} Clade specific model - diversity dependent
\item \code{"rr_k_0lama"} Clade specific model - diversity dependent
}}
}
\value{
A named list with \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}} arguments.
}
\description{
Set up DAISIE_ML arguments
}
\examples{
model <- "cr_dd"
setup_model(
  model = model
)
}
\author{
Luis M Valente, Pedro Neves, Joshua W. Lambert
}
