% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model.R
\name{setup_model}
\alias{setup_model}
\title{Set up DAISIE_ML arguments}
\usage{
setup_model(model, low_rates = FALSE)
}
\arguments{
\item{model}{A string with model that should run. Models are as follows:
\itemize{
\item \code{"cr_dd"} Clade specific model - diversity dependent. All parameters
free.
\item \code{"cr_di"} Clade specific model - diversity independent (K = Inf). All
other parameters free.
\item \code{"cr_dd_0laa"} Clade specific model - diversity dependent and no
anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0laa"} Clade specific model - diversity independent (K = Inf)
and no anagenesis (laa fixed to zero). All other parameters free.
\item \code{"cr_di_0lac"} Clade specific model - diversity indendent (K = Inf)
and no cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"cr_dd_0lac"} Clade specific model - diversity dependent and no
cladogenesis (lac fixed to zero). All other parameters free.
\item \code{"rr_lac_di"} Clade specific model - diversity independent (K = Inf)
and relaxed cladogenesis. All other parameters free.
\item \code{"rr_lac_dd"} Clade specific model - diversity dependent and relaxed
cladogenesis. All other parameters free.
\item \code{"rr_mu_di"} Clade specific model - diversity independent (K = Inf) and
relaxed extinction. All other parameters free.
\item \code{"rr_mu_dd"} Clade specific model - diversity dependent and relaxed
extinction. All other parameters free.
\item \code{"rr_k"} Clade specific model - diversity dependent and relaxed carrying
capacity. All other parameters free.
\item \code{"rr_gam_di"} Clade specific model - diversity independent (K = Inf)
and relaxed colonisation. All other parameters free.
\item \code{"rr_gam_dd"} Clade specific model - diversity dependent and relaxed
colonisation. All other parameters free.
\item \code{"rr_laa_di"} Clade specific model - diversity independent (K = Inf)
and relaxed anagenesis. All other parameters free.
\item \code{"rr_laa_dd"} Clade specific model - diversity dependent and relaxed
anagenesis. All other parameters free.
\item \code{"rr_mu_di_0lac"} Clade specific model - diversity independent,
relaxed extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_mu_dd_0lac"} Clade specific model - diversity dependent, relaxed
extinction, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_k_0lac"} Clade specific model - diversity dependent, relaxed
carrying capacity, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_gam_di_0lac"} Clade specific model - diversity independent,
relaxed colonisation, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"r_gam_dd_0lac"} Clade specific model - diversity dependent, relaxed
colonisation, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_laa_di_0lac"} Clade specific model - diversity independent,
relaxed anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"r_laa_dd_0lac"} Clade specific model - diversity dependent, relaxed
anagenesis, and no cladogenesis (lac fixed to zero). All other
parameters free.
\item \code{"rr_lac_di_0laa"} Clade specific model - diversity independent,
relaxed cladogenesis, and no anagenesis (laa fixed to zero)
\item \code{"rr_lac_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_mu_di_0laa"} Clade specific model - diversity independent
\item \code{"rr_mu_dd_0laa"} Clade specific model - diversity dependent
\item \code{"rr_k_0laa"} Clade specific model - diversity dependent
\item \code{"rr_gam_di_0laa"} Clade specific model - diversity independent, relaxed
colonisation, and no anagenesis (laa fixed to zero). All other parameters
free
\item \code{"rr_gam_dd_0laa"} Clade specific model - diversity dependent, relaxed
colonisation, and no anagenesis (laa fixed to zero). All other parameters
free
}}

\item{low_rates}{Boolean determining whether the random sampling of initial
parameter estimates should be sampled from a broad range (FALSE) or from a
restricted range where the initial rates a ensured to be smaller (TRUE).
The latter helps when using large datasets that may fail the initial
likelihood computation with higher rates that could be sampled from the
broad range of rates.}
}
\value{
A named list with \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}} arguments.
}
\description{
Set up DAISIE_ML arguments
}
\examples{
model <- "cr_dd"
setup_model(
  model = model
)
}
\author{
Luis M Valente, Pedro Santos Neves, Joshua W. Lambert
}
