% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_std_pars2.R
\name{setup_std_pars2}
\alias{setup_std_pars2}
\title{Fills pars2 vector for most common DAISIE ML CS runs}
\usage{
setup_std_pars2(
  res = 100,
  ddmodel,
  cond,
  verbose = TRUE,
  island_ontogeny = 0,
  eqmodel = 0,
  maxiter = 1000 * round((1.25)^5),
  tol = c(1e-04, 1e-05, 1e-07),
  x_E = 0.95,
  x_I = 0.95
)
}
\arguments{
\item{cond}{An integer specifying conditioning, as described in
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML_CS()}}.}
}
\value{
A numeric vector of length 12 containing pars2 for
\code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}}
}
\description{
Returns the pars2 vector needed for the most simple \code{DAISIE_ML_CS()} runs,
to facilitate computing the loglikelihood in these cases. This function
should only be assumed correct for the LL setup of the CS model, 1 type,
constant rate no equilibrium models. Tolerances and other technical
parameters need not be specified, as the returned values match the default
values of \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}}. The only arguments without default values
are \code{ddmodel} and \code{cond}, as these are most often varied when checking model
output. Regardless, all other default parameter values can be forced to
meet specific needs of more complex models.
}
\author{
Pedro Santos Neves
}
