% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_goodness_of_fit.R
\name{summarize_bootstrap_results}
\alias{summarize_bootstrap_results}
\title{Computes summary results of bootstrap simulations useful for plotting}
\usage{
summarize_bootstrap_results(
  simulation_dataset,
  empirical_summary_statistics,
  m_pool = 1000
)
}
\arguments{
\item{simulation_dataset}{output of DAISIE_sim}

\item{empirical_summary_statistics}{vector with the number of species, number
of colonization, size of the largest clade and the rank of the largest clade
in the empirical data}

\item{m_pool}{size of the mainland pool}

\item{time}{age of the island or archipelago}
}
\value{
overall_results a list of results
}
\description{
The output is a list of results
}
\author{
Rampal S. Etienne & Luis Valente
}
