% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_goodness_of_fit.R
\name{summarize_bootstrap_results}
\alias{summarize_bootstrap_results}
\title{Computes summary results of bootstrap simulations useful for plotting}
\usage{
summarize_bootstrap_results(daisie_data, mainland_n = 1000)
}
\arguments{
\item{daisie_data}{A list, conforming to the \code{\link{DAISIE}} object format. Usually
preprocessed by \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}, see it's documentation for
details. Otherwise it may be generated via simulations, using
\code{\link[DAISIE:DAISIE_sim]{DAISIE::DAISIE_sim_cr()}} and friends.}

\item{mainland_n}{A numeric stating the number of mainland species, that
is the number of species that can potentially colonize the island.
If using a clade-specific diversity dependence, this value is set to 1
internally in the simulation.
If using an island-wide diversity dependence, this value is set to the
number of mainland species.}
}
\value{
overall_results a list of results
}
\description{
The output is a list of results
}
\examples{
\dontrun{
clado_rate <- 0.5
ext_rate <- 0.2
carr_cap <- Inf
immig_rate <- 0.005
ana_rate <- 1
sim_pars <- c(clado_rate, ext_rate, carr_cap, immig_rate, ana_rate)
set.seed(1)

dataset_cs <- DAISIE::DAISIE_sim_cr(
  time = 10,
  M = 1000,
  pars = sim_pars,
  replicates = 10,
  plot_sims = FALSE,
  verbose = FALSE,
  divdepmodel = "CS"
)

dataset_iw <- DAISIE::DAISIE_sim_cr(
  time = 10,
  M = 1000,
  pars = sim_pars,
  replicates = 10,
  plot_sims = FALSE,
  verbose = FALSE,
  divdepmodel = "IW"
)
overall_results_cs <- DAISIEutils::summarize_bootstrap_results(
  daisie_data = dataset_cs
)
overall_results_iw <- DAISIEutils::summarize_bootstrap_results(
  daisie_data = dataset_iw
)
}
}
\author{
Rampal S. Etienne & Luis Valente
}
